requirements.txt
python-dotenv
pandas
pandasai
langchain_groq
streamlit